Go crazy generating new tours
The old way of generating tours is O(n) with some unnecessary copying.

Let's turn that into O(1).

We can't make copies if we want it to be O(1), so we have to mutate the array in place.
We also have to do a bunch of other crazy stuff.
All in the name of performance.